small update

+++ Sugar/Sugar.Language/Parsing/Nodes/DataTypes/Enums/CreationType.cs
@@ -4,7 +4,7 @@ using Sugar.Language.Analysis.ProjectStructure.Enums;
 
 namespace Sugar.Language.Parsing.Nodes.DataTypes.Enums
 {
-    internal enum CreationType : byte
+    internal enum CreationType : ushort
     {
         Enum = ProjectMemberEnum.Enum,
         Class = ProjectMemberEnum.Class,
